+++ docs/examples/CarouselUncontrolled.js
@@ -1,26 +1,26 @@
 const carouselInstance = (
   <Carousel>
-    <CarouselItem>
+    <Carousel.Item>
       <img width={900} height={500} alt="900x500" src="/assets/carousel.png"/>
-      <div className="carousel-caption">
+      <Carousel.Caption>
         <h3>First slide label</h3>
         <p>Nulla vitae elit libero, a pharetra augue mollis interdum.</p>
-      </div>
-    </CarouselItem>
-    <CarouselItem>
+      </Carousel.Caption>
+    </Carousel.Item>
+    <Carousel.Item>
       <img width={900} height={500} alt="900x500" src="/assets/carousel.png"/>
-      <div className="carousel-caption">
+      <Carousel.Caption>
         <h3>Second slide label</h3>
         <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>
-      </div>
-    </CarouselItem>
-    <CarouselItem>
+      </Carousel.Caption>
+    </Carousel.Item>
+    <Carousel.Item>
       <img width={900} height={500} alt="900x500" src="/assets/carousel.png"/>
-      <div className="carousel-caption">
+      <Carousel.Caption>
         <h3>Third slide label</h3>
         <p>Praesent commodo cursus magna, vel scelerisque nisl consectetur.</p>
-      </div>
-    </CarouselItem>
+      </Carousel.Caption>
+    </Carousel.Item>
   </Carousel>
 );