+++ docs/src/sections/BreadcrumbSection.js
@@ -9,10 +9,10 @@ export default function BreadcrumbSection() {
   return (
     <div className="bs-docs-section">
       <h2 className="page-header">
-        <Anchor id="breadcrumbs">Breadcrumbs</Anchor> <small>Breadcrumb, BreadcrumbItem</small>
+        <Anchor id="breadcrumbs">Breadcrumbs</Anchor> <small>Breadcrumb, Breadcrumb.Item</small>
       </h2>
 
-      <p>Breadcrumbs are used to indicate the current page's location. Add <code>active</code> attribute to active <code>BreadcrumbItem</code>.</p>
+      <p>Breadcrumbs are used to indicate the current page's location. Add <code>active</code> attribute to active <code>Breadcrumb.Item</code>.</p>
       <p>Do not set both <code>active</code> and <code>href</code> attributes. <code>active</code> overrides <code>href</code> and <code>span</code> element is rendered instead of <code>a</code>.</p>
 
       <h3><Anchor id="breadcrumbs-example">Breadcrumbs Example</Anchor></h3>
@@ -21,7 +21,7 @@ export default function BreadcrumbSection() {
       <h3><Anchor id="breadcrumbs-props">Props</Anchor></h3>
       <p><code>Breadcrumb</code> component itself doesn't have any specific public properties</p>
 
-      <h4><Anchor id="breadcrumbs-props-breadcrumbItem">BreadcrumbItem</Anchor></h4>
+      <h4><Anchor id="breadcrumbs-props-breadcrumb-item">Breadcrumb.Item</Anchor></h4>
       <PropTable component="BreadcrumbItem"/>
     </div>
   );